requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
yt-dlp
ffmpeg-python
python-multipart
Markdown